requirements.txt
apparmor==3.0.4
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.1
btrfsutil==5.18.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cupshelpers==1.0
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
Django==4.0.4
django-cors-headers==3.12.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
docutils==0.18.1
evdev==1.5.0
file-magic==0.4.0
filelock==3.7.1
html5lib==1.1
idna==3.3
itypes==1.2.0
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
Jinja2==3.1.2
keyutils==0.6
LibAppArmor==3.0.4
libtorrent==2.0.6
lit==13.0.1.dev0
louis==3.22.0
lxml==4.9.0
manjaro-sdk==0.1
MarkupSafe==2.1.1
meson==0.62.2
more-itertools==8.13.0
msgpack==1.0.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
oauthlib==3.2.0
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pep517==0.12.0
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
ply==3.11
progress==1.6
psutil==5.9.1
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
PyGObject==3.42.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
pysmbc==1.0.23
python-xlib==0.31
python3-openid==3.2.0
pytz==2022.1
pyudev==0.22.0.dev20211202
pyxdg==0.28
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.10
requests==2.27.1
requests-oauthlib==1.3.1
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
setproctitle==1.2.3
shiboken6==6.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
solaar==1.1.3
sqlparse==0.4.2
systemd-python==234
team==1.0
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.6.23
udiskie==2.4.2
uritemplate==4.1.1
urllib3==1.26.9
validate-email==1.3
validate-pyproject==0.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
webencodings==0.5.1
